# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Neel Kamal Homestay API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /media:
    get:
      operationId: listMedia
      tags: [media]
      summary: List all media (admin)
      parameters:
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: usedIn
          schema:
            type: string
            enum: [hero, gallery, room, about, footer]
      responses:
        "200":
          description: Media list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Media"
  /media/upload:
    post:
      operationId: uploadMedia
      tags: [media]
      summary: Upload a media file (admin)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/MediaUploadForm"
      responses:
        "201":
          description: Uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Media"
  /media/{id}:
    get:
      operationId: getMedia
      tags: [media]
      summary: Get a media item
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Media item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Media"

    put:
      operationId: updateMedia
      tags: [media]
      summary: Update media metadata (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MediaUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Media"

    delete:
      operationId: deleteMedia
      tags: [media]
      summary: Delete a media file (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /media/{id}/usage:
    patch:
      operationId: updateMediaUsage
      tags: [media]
      summary: Set usage context for a media item (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MediaUsageInput"
      responses:
        "200":
          description: Usage updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Media"

  # ─── GALLERY ─────────────────────────────────────────────────────────────────
...
  /reviews:
    get:
      operationId: listReviews
      tags: [reviews]
      summary: List reviews
      parameters:
        - in: query
          name: admin
          schema:
            type: string
          description: If "true", return all including hidden
      responses:
        "200":
          description: Reviews list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"

    post:
      operationId: createReview
      tags: [reviews]
      summary: Create a review (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
  /reviews/{id}:
    put:
      operationId: updateReview
      tags: [reviews]
      summary: Update a review (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"

    delete:
      operationId: deleteReview
      tags: [reviews]
      summary: Delete a review (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /reviews/{id}/toggle-visible:
    patch:
      operationId: toggleReviewVisible
      tags: [reviews]
      summary: Toggle review visibility (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"

  # ─── BOOKINGS ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Media:
      type: object
      properties:
        id:
          type: integer
        filename:
          type: string
        originalName:
          type: string
        url:
          type: string
        mimeType:
          type: string
        sizeBytes:
          type: integer
        width:
          type: ["integer", "null"]
        height:
          type: ["integer", "null"]
        altText:
          type: ["string", "null"]
        usedIn:
          type: array
          items:
            type: string
        createdAt:
          type: string
      required: [id, filename, originalName, url, mimeType, sizeBytes, usedIn, createdAt]
    MediaUploadForm:
      type: object
      properties:
        file:
          type: string
          format: binary
        altText:
          type: string
      required: [file]
    MediaUpdate:
      type: object
      properties:
        altText:
          type: string
        originalName:
          type: string
    MediaUsageInput:
      type: object
      properties:
        usedIn:
          type: array
          items:
            type: string
      required: [usedIn]

    # ── Gallery ───────────────────────────────────────────────────────────────
...
    Review:
      type: object
      properties:
        id:
          type: integer
        guestName:
          type: string
        rating:
          type: integer
        reviewText:
          type: string
        stayDate:
          type: ["string", "null"]
        source:
          type: ["string", "null"]
        isVisible:
          type: boolean
        createdAt:
          type: string
      required: [id, guestName, rating, reviewText, isVisible, createdAt]
    ReviewInput:
      type: object
      properties:
        guestName:
          type: string
        rating:
          type: integer
        reviewText:
          type: string
        stayDate:
          type: string
        source:
          type: string
        isVisible:
          type: boolean
      required: [guestName, rating, reviewText]

    # ── Bookings ──────────────────────────────────────────────────────────────
...